requirements.txt
Flask
Werkzeug
tensorflow>=2.16.1
numpy
pandas
scikit-learn
requests
python-dotenv
pillow
matplotlib
seaborn
flask-sqlalchemy
SQLAlchemy
flask-wtf
pytest
gunicorn
opencv-python
python-crontab
flask-login 